Make a clickSelects geom_tallrect that completely tiles the x range. This makes it easy to construct tallrects for the common case of selecting a particular x value.

Make a clickSelects geom_tallrect that completely tiles the x range. This makes it easy to construct tallrects for the common case of selecting a particular x value.

Usage

make_tallrect(data, x.name, even = FALSE, alpha = 1/2, ...)

Arguments

data

data.frame to analyze for unique x.name values.

x.name

variable to be used for x, clickSelects.

even

Logical parameter, should tallrects be of even width?

alpha

transparency of a selected tallrect, default 1/2.

...

passed to geom_tallrect.

Value

a geom_tallrect layer.

Author

Toby Dylan Hocking